+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-- Added "Auto Selected" skin to the default skin pack. Selecting this skin will reset your skin to its default appearance. (Legacy4J only)
-- Removed Reset Skin keybind, as it is no longer needed. (Legacy4J only)
+- Added skin pack icons to the change skin screen (Legacy4J only)
+- Improved the arm animations for skins using `animationArmsOutFront`
+- Added single leg animation support for skins using `animationSingleLegAnimation` 
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master